adding in compute
