--- azure/rs_azure_lb/changelog.md
@@ -1,5 +1,0 @@
-LB Plugin changelog
-
-v1.0
------
-- initial release
